Setting up FSM to drive PID
there are some things that need to be checked. such as that value to drive the analog output to when you hit the limit and how to determine you are going out then.
There needs to be a way of deciding when to start a cycle, and more code for things like emergency stop